refactor: Extract cold compaction code into a module like hot
