+++ compactor/src/handler.rs
@@ -1,24 +1,20 @@
 //! Compactor handler
 
+use crate::{cold, compact::Compactor, hot};
 use async_trait::async_trait;
-use backoff::Backoff;
 use futures::{
     future::{BoxFuture, Shared},
-    FutureExt, StreamExt, TryFutureExt,
+    FutureExt, TryFutureExt,
 };
 use iox_query::exec::Executor;
-use metric::Attributes;
 use observability_deps::tracing::*;
 use std::sync::Arc;
-
 use thiserror::Error;
 use tokio::{
     task::{JoinError, JoinHandle},
     time::Duration,
 };
 use tokio_util::sync::CancellationToken;
-
-use crate::{compact::Compactor, hot};
 
 #[derive(Debug, Error)]
 #[allow(missing_copy_implementations, missing_docs)]
@@ -185,7 +181,7 @@ pub async fn run_compactor_once(compactor: Arc<Compactor>) {
         }
     }
     debug!("start cold cycle");
-    compacted_partitions += compact_cold_partitions(Arc::clone(&compactor)).await;
+    compacted_partitions += cold::compact(Arc::clone(&compactor)).await;
 
     if compacted_partitions == 0 {
         // sleep for a second to avoid a busy loop when the catalog is polled
@@ -196,121 +192,6 @@ pub async fn run_compactor_once(compactor: Arc<Compactor>) {
         num_cold_cycles = 1,
         "complete running compactor once that includes"
     );
-}
-
-async fn compact_cold_partitions(compactor: Arc<Compactor>) -> usize {
-    let cold_attributes = Attributes::from(&[("partition_type", "cold")]);
-    // Select cold partition candidates
-    let start_time = compactor.time_provider.now();
-    let candidates = Backoff::new(&compactor.backoff_config)
-        .retry_all_errors("cold_partitions_to_compact", || async {
-            compactor
-                .cold_partitions_to_compact(compactor.config.max_number_partitions_per_shard)
-                .await
-        })
-        .await
-        .expect("retry forever");
-    if let Some(delta) = compactor
-        .time_provider
-        .now()
-        .checked_duration_since(start_time)
-    {
-        let duration = compactor
-            .candidate_selection_duration
-            .recorder(cold_attributes.clone());
-        duration.record(delta);
-    }
-
-    // Add other compaction-needed info into selected partitions
-    let start_time = compactor.time_provider.now();
-    let candidates = Backoff::new(&compactor.backoff_config)
-        .retry_all_errors("add_info_to_partitions", || async {
-            compactor.add_info_to_partitions(&candidates).await
-        })
-        .await
-        .expect("retry forever");
-    if let Some(delta) = compactor
-        .time_provider
-        .now()
-        .checked_duration_since(start_time)
-    {
-        let duration = compactor
-            .partitions_extra_info_reading_duration
-            .recorder(cold_attributes.clone());
-        duration.record(delta);
-    }
-
-    let n_candidates = candidates.len();
-    if n_candidates == 0 {
-        debug!("no cold compaction candidates found");
-        return 0;
-    } else {
-        debug!(n_candidates, "found cold compaction candidates");
-    }
-
-    let start_time = compactor.time_provider.now();
-
-    // Repeat compacting n cold partitions in parallel until all candidates are compacted.
-    // Concurrency level calculation (this is estimated from previous experiments. The actual
-    // resource management will be more complicated and a future feature):
-    //
-    //   . Each `compact partititon` takes max of this much memory cold_input_size_threshold_bytes
-    //   . We have this memory budget: max_cold_concurrent_size_bytes
-    // --> num_parallel_partitions = max_cold_concurrent_size_bytes/
-    //     cold_input_size_threshold_bytes
-    let num_parallel_partitions = (compactor.config.max_cold_concurrent_size_bytes
-        / compactor.config.cold_input_size_threshold_bytes)
-        as usize;
-
-    futures::stream::iter(candidates)
-        .map(|p| {
-            // run compaction in its own task
-            let comp = Arc::clone(&compactor);
-            tokio::task::spawn(async move {
-                let partition_id = p.candidate.partition_id;
-                let p = Arc::new(p);
-                let compaction_result =
-                    crate::compact_cold_partition(&comp, p, &Default::default()).await;
-
-                match compaction_result {
-                    Err(e) => {
-                        warn!(?e, ?partition_id, "cold compaction failed");
-                    }
-                    Ok(_) => {
-                        debug!(?partition_id, "cold compaction complete");
-                    }
-                };
-            })
-        })
-        // Assume we have enough resources to run
-        // num_parallel_partitions compactions in parallel
-        .buffer_unordered(num_parallel_partitions)
-        // report any JoinErrors (aka task panics)
-        .map(|join_result| {
-            if let Err(e) = join_result {
-                warn!(?e, "cold compaction task failed");
-            }
-            Ok(())
-        })
-        // Errors are reported during execution, so ignore results here
-        // https://stackoverflow.com/questions/64443085/how-to-run-stream-to-completion-in-rust-using-combinators-other-than-for-each
-        .forward(futures::sink::drain())
-        .await
-        .ok();
-
-    // Done compacting all candidates in the cycle, record its time
-    if let Some(delta) = compactor
-        .time_provider
-        .now()
-        .checked_duration_since(start_time)
-    {
-        let duration = compactor
-            .compaction_cycle_duration
-            .recorder(cold_attributes);
-        duration.record(delta);
-    }
-
-    n_candidates
 }
 
 #[async_trait]
